+++ packages/browser-wallet/src/popup/pages/Account/i18n/en.ts
@@ -5,6 +5,7 @@ const t = {
     accountAddress: 'Account address',
     siteConnected: 'Connected',
     siteNotConnected: 'Not connected',
+    accountBalanceError: 'Unable to retrieve account balance',
     actions: {
         log: 'transaction log',
         send: 'send ccd',
@@ -30,6 +31,21 @@ const t = {
             show: 'Show private key',
             done: 'Done',
         },
+    },
+    sendCcd: {
+        labels: {
+            ccd: 'Enter amount to transfer',
+            recipient: 'Enter recipient address',
+        },
+        buttons: {
+            back: 'Back',
+            send: 'Send',
+            finish: 'Finish',
+            continue: 'Continue',
+        },
+        title: 'Send CCD',
+        receiptTitle: 'Transaction: Send CCD',
+        fee: 'Estimated transaction fee',
     },
     accountPending: 'This account is still pending finalization.',
 };