+++ packages/browser-wallet/src/popup/pages/Setup/Setup.tsx
@@ -27,7 +27,7 @@ function Intro() {
             </div>
             <Button
                 className="onboarding-setup__intro-wrapper__continue-button"
-                width="narrow"
+                width="medium"
                 onClick={() => navigate(setupRoutes.passcode)}
             >
                 {t('continue')}